+++ lib/src/json.dart
@@ -1,4 +1,63 @@
+import 'dart:convert' as dart_convert;
+
+import 'package:swiss_knife/src/string.dart';
+import 'package:swiss_knife/swiss_knife.dart';
+
 import 'collections.dart';
+
+/// Parses [json].
+///
+/// [def] The default value if [json] is null, empty or blank String.
+dynamic parseJSON(dynamic json, [dynamic def]) {
+  if (json == null) return def;
+
+  if (json is String) {
+    if (json.isEmpty || (json.length < 100 && isBlankString(json))) return def;
+    return dart_convert.json.decode(json);
+  } else {
+    return json;
+  }
+}
+
+/// Encodes [json].
+String encodeJSON(dynamic json,
+    {String ident, bool withIdent, bool clearNullEntries}) {
+  clearNullEntries ??= false;
+  if (clearNullEntries) {
+    removeNullEntries(json);
+  }
+
+  if (withIdent != null && withIdent) {
+    if (ident == null || ident.isEmpty) {
+      ident = '  ';
+    }
+  }
+
+  dart_convert.JsonEncoder encoder;
+
+  if (ident != null && ident.isNotEmpty) {
+    encoder = dart_convert.JsonEncoder.withIndent(ident);
+  } else {
+    encoder = dart_convert.JsonEncoder();
+  }
+
+  return encoder.convert(json);
+}
+
+/// Remove null entries from [json] tree.
+T removeNullEntries<T>(T json) {
+  if (json == null) return json;
+
+  if (json is List) {
+    json.removeWhere((e) => null);
+    json.forEach(removeNullEntries);
+  } else if (json is Map) {
+    json.removeWhere((key, value) => key == null || value == null);
+    json.values.forEach(removeNullEntries);
+  }
+
+  return json;
+}
 
 /// Returns [true] if [value] is a JSON.
 bool isJSON(dynamic value) {